fixed headerlayout and media queries
