+++ src/components/Layout/Header/Searchbar/Searchbar.jsx
@@ -1,6 +1,6 @@
 import { useState } from "react";
 import { useNavigate } from "react-router-dom";
-import { APIUrl } from "../../../../api/api";
+import { APIUrl } from "../../../../hooks/useApi";
 import PropTypes from "prop-types";
 
 export default function Searchbar({ className }) {
@@ -37,9 +37,9 @@ export default function Searchbar({ className }) {
     };
 
     return (
-        <div className={`relative ${className}`} >
-            <form onSubmit={handleSearch} className="flex items-center px-3">
-                <div className=" flex items-center border bg-white/50 border-gray-300 rounded-1-md p-2">
+        <div className={`relative w-full md:w-fit ${className}`} >
+            <form onSubmit={handleSearch} className="flex justify-center md:justify-end items-center md:px-3">
+                <div className=" flex items-center w-full border bg-white/50 border-gray-300 rounded p-2">
                     <svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" fill="currentColor" className="bi bi-search me-1" viewBox="0 0 16 16">
                          <path d="M11.742 10.344a6.5 6.5 0 1 0-1.397 1.398h-.001q.044.06.098.115l3.85 3.85a1 1 0 0 0 1.415-1.414l-3.85-3.85a1 1 0 0 0-.115-.1zM12 6.5a5.5 5.5 0 1 1-11 0 5.5 5.5 0 0 1 11 0"/>
                     </svg>
